Extract logic to PeerDiscoverer
