+++ packages/contracts/source/contracts/p2p/index.ts
@@ -10,6 +10,7 @@ export * from "./network-state";
 export * from "./peer";
 export * from "./peer-communicator";
 export * from "./peer-connector";
+export * from "./peer-discoverer";
 export * from "./peer-processor";
 export * from "./peer-repository";
 export * from "./peer-verifier";
