+++ packages/contracts/source/contracts/p2p/peer-discoverer.ts
@@ -1,0 +1,3 @@
+export interface PeerDiscoverer {
+	populateSeedPeers(): Promise<void>;
+}
